pyproject.toml
[project]
name = "k8s-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.44.1",
    "openai-agents>=0.2.11",
    "python-dotenv>=1.1.1",
]
